requirements.txt
python-dotenv
fastapi 
uvicorn 
boto3 
PyJWT 
requests 
networkx 
pydantic 
matplotlib
python-multipart